Payment

+++ bookStoreApi/store/views.py
@@ -1,5 +1,5 @@
 
-from .serializers import BookOnSaleSerializer,OrderDetailSerializer,OrderSerializer
+from .serializers import BookOnSaleSerializer,OrderSerializer,CheckoutSerializer
 from rest_framework import status,permissions
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -8,6 +8,9 @@ from .models import BookOnSale,OrderDetail,Order
 from book.serializers import BookSerializer
 from core.mixins import IsBookExist,IsSaleExist
 from core.custom_exceptions import CustomAPIException
+from core.mollie import createMolliePayment
+from django.db import transaction
+
 class SellBookView(APIView):
     parser_classes = (MultiPartParser, FormParser)  # To handle file uploads
     permission_classes=[permissions.IsAuthenticated]
@@ -66,24 +69,45 @@ class CheckOutView(APIView):
     permission_classes=[permissions.IsAuthenticated]
     
     def put(self,request,*args,**kwargs):
-        open_order,open_order_created=Order.objects.get_or_create(customer=request.user,status="OPEN")
-        if open_order_created or open_order.cost==0:
-            raise CustomAPIException("Your cart is empty! You can not continue to checkout.",status=400)
-        deleted_detail=0
-        for order_detail in open_order.order_details.all():
-            if order_detail.book_on_sale.status != "OPEN":
-                order_detail.delete()
-                deleted_detail+=1
-        if deleted_detail>0:
+        checkoutSerializer=CheckoutSerializer(data=request.data)
+        if checkoutSerializer.is_valid():
+            open_order,open_order_created=Order.objects.get_or_create(customer=request.user,status="OPEN")
+            if open_order_created or open_order.cost==0:
+                raise CustomAPIException("Your cart is empty! You can not continue to checkout.",status=400)
+            
+            deleted_details=[]
+            order_details=open_order.order_details.all()
+
+            for order_detail in order_details:
+                if order_detail.book_on_sale.status != "OPEN":
+                    order_detail.delete()
+                    deleted_details.append(order_detail.book_on_sale.book.name)
+
+            if len(deleted_details)>0:
+                open_order.refresh_from_db()
+                serializer=OrderSerializer(open_order)
+                raise CustomAPIException("The book(s) you have added to your cart is not available anymore! {}".format(*deleted_details),status=404,data=serializer.data)
+            try:
+
+                with transaction.atomic():
+                    for order_detail in order_details:
+                        order_detail.book_on_sale.status="PENDING"
+                        order_detail.book_on_sale.save()
+                    
+                    open_order.status="PENDING"
+                    open_order.save()
+                    payment=createMolliePayment(str(open_order.cost),open_order.id,checkoutSerializer.validated_data.get('redirectUrl'))
+            except:
+                open_order.refresh_from_db()
+                raise CustomAPIException("Checkout can not process, please try again later.", 500, data=open_order)
+
+            
             open_order.refresh_from_db()
+
+            
             serializer=OrderSerializer(open_order)
-            raise CustomAPIException("The book(s) you have added to your cart is not available anymore! {}".format(order_detail.book_on_sale.book.name),status=404,data=serializer.data)
-        
-        for order_detail in open_order.order_details.all():
-            order_detail.book_on_sale.status="PENDING"
-            order_detail.book_on_sale.save()
-        open_order.status="PENDING"
-        open_order.save()
-        open_order.refresh_from_db()
-        serializer=OrderSerializer(open_order)
-        return Response(serializer.data,status=status.HTTP_200_OK)
+            response_data=serializer.data
+            response_data["redirectUrl"]=payment["_links"]["checkout"]["href"]
+
+            return Response(response_data,status=status.HTTP_200_OK)
+        raise CustomAPIException(str(checkoutSerializer.errors),status=400)
